+++ src/main/java/com/sarel/web/service/ExpedienteLaboratorioService.java
@@ -1,5 +1,6 @@
 package com.sarel.web.service;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -19,7 +20,7 @@ public interface ExpedienteLaboratorioService {
 	
 	List<ExpedienteLaboratorio> findVeinte();
 
-	ExpedienteLaboratorio findByCarnet(Integer carne);
+	ExpedienteLaboratorio findByCarnet(BigInteger carne);
 	
 	ExpedienteLaboratorio findByIdPaciente(Integer idPaciente);
